feat: added temporary layout for task status columns, organised into folder
